+++ setup.py
@@ -31,6 +31,7 @@ packages = ['mesonbuild',
             'mesonbuild.backend',
             'mesonbuild.cmake',
             'mesonbuild.compilers',
+            'mesonbuild.compilers.mixins',
             'mesonbuild.dependencies',
             'mesonbuild.modules',
             'mesonbuild.scripts',
